requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
pip install wikipedia
pip install -U ddgs
